Fix issues with incremental updates using different tags

+++ lib/right_scraper/scrapers/git_scraper.rb
@@ -48,24 +48,27 @@ module RightScale
       msg += "git repository '#{@repo.display_name}'"
       @callback.call(msg, is_step=true) if @callback
       ssh_cmd = ssh_command
-      is_tag = nil
-      is_branch = nil
+      is_tag  = is_branch = on_branch = nil
+      has_tag = !@repo.tag.nil? && !@repo.tag.empty?
 
       if @incremental
         Dir.chdir(@current_repo_dir) do
           analysis = git_fetch_and_analyze(ssh_cmd, update=true)
-          if succeeded? && @incremental
+          if succeeded? && @incremental && has_tag
             is_tag = analysis[:tag]
             is_branch = analysis[:branch]
+            on_branch = analysis[:on_branch]
             if !is_tag && !is_branch
-              @callback.call('Nothing to update: repo tag refers to neither a branch nor a tag', is_step=false) if @callback
-              return true
+              current_sha = `git rev-parse HEAD`.chomp
+              if current_sha == @repo.tag
+                @callback.call("Nothing to update: already using #{repo.tag}", is_step=false) if @callback
+                return true
+              end
             end
             if is_tag && is_branch
               @errors << 'Repository tag ambiguous: could be git tag or git branch'
-            else
-              tag = @repo.tag.nil? || @repo.tag.empty? ? 'master' : @repo.tag
-              res = `git checkout #{tag} 2>&1`
+            elsif !on_branch
+              res = `git checkout #{@repo.tag} 2>&1`
               if $? != 0
                 @callback.call("Failed to checkout #{tag}: #{res}, falling back to cloning", is_step=false) if @callback
                 FileUtils.rm_rf(@current_repo_dir)
@@ -79,24 +82,27 @@ module RightScale
         git_cmd = "#{ssh_cmd} git clone --quiet --depth 1 \"#{@repo.url}\" \"#{@current_repo_dir}\" 2>&1"
         res = @watcher.launch_and_watch(git_cmd, @current_repo_dir)
         handle_watcher_result(res, 'git clone', update=false)
-        if !@repo.tag.nil? && !@repo.tag.empty? && @repo.tag != 'master' && succeeded?
+        if has_tag && succeeded?
           Dir.chdir(@current_repo_dir) do
             if is_tag.nil?
-              analysis = git_fetch_and_analyze(ssh_cmd, update=false)
-              is_tag = analysis[:tag]
+              analysis  = git_fetch_and_analyze(ssh_cmd, update=false)
+              is_tag    = analysis[:tag]
               is_branch = analysis[:branch]
+              on_branch = analysis[:on_branch]
             end
             if succeeded?
               if is_tag && is_branch
                 @errors << 'Repository tag ambiguous: could be git tag or git branch'
-              elsif is_branch
-                output = `git branch #{@repo.tag} origin/#{@repo.tag} 2>&1`
-                @errors << output if $? != 0
+              elsif is_branch 
+                if !on_branch
+                  output = `git branch #{@repo.tag} origin/#{@repo.tag} 2>&1`
+                  @errors << output if $? != 0
+                end
               elsif !is_tag # Not a branch nor a tag, SHA ref? fetch everything so we have all SHAs
-                output = `#{ssh_cmd} git fetch origin master --depth #{2**31 - 1} 2>&1`
+                output = `#{ssh_cmd} git fetch --depth #{2**31 - 1} 2>&1`
                 @errors << output if $? != 0
               end
-              if succeeded?
+              if succeeded? && !on_branch
                 output = `git checkout #{@repo.tag} 2>&1`
                 @errors << output if $? != 0
               end
@@ -163,27 +169,27 @@ module RightScale
     # === Raise
     # Exception:: If the USERPROFILE environment variable is not set
     def win32_ssh_command
-	  key_content = @repo.first_credential
-	  unless key_content.nil?
-		# resolve key file path.
-		raise 'Environment variable USERPROFILE is missing' unless ENV['USERPROFILE']
-		user_profile_dir_path = ENV['USERPROFILE']
-		ssh_keys_dir = File.join(user_profile_dir_path, '.ssh')
-		FileUtils.mkdir_p(ssh_keys_dir) unless File.directory?(ssh_keys_dir)
-		ssh_key_file_path = File.join(ssh_keys_dir, 'id_rsa')
-
-		# (re)create key file. must overwrite any existing credentials in case
-		# we are switching repositories and have different credentials for each.
-		File.open(ssh_key_file_path, 'w') { |f| f.puts(key_content) }
-
-		# we need to create the "known_hosts" file or else the process will
-		# halt in windows waiting for a yes/no response to the unknown
-		# git host. this is normally handled by specifying
-		# "-o StrictHostKeyChecking=no" in the GIT_SSH executable, but it is
-		# still a mystery why this doesn't work properly in windows.
-		# so make a ssh call which creates the proper "known_hosts" file.
-		system("ssh -o StrictHostKeyChecking=no #{repo.url.split(':').first} exit 2>&1")
-	  end
+      key_content = @repo.first_credential
+      unless key_content.nil?
+        # resolve key file path.
+        raise 'Environment variable USERPROFILE is missing' unless ENV['USERPROFILE']
+        user_profile_dir_path = ENV['USERPROFILE']
+        ssh_keys_dir = File.join(user_profile_dir_path, '.ssh')
+        FileUtils.mkdir_p(ssh_keys_dir) unless File.directory?(ssh_keys_dir)
+        ssh_key_file_path = File.join(ssh_keys_dir, 'id_rsa')
+    
+        # (re)create key file. must overwrite any existing credentials in case
+        # we are switching repositories and have different credentials for each.
+        File.open(ssh_key_file_path, 'w') { |f| f.puts(key_content) }
+    
+        # we need to create the "known_hosts" file or else the process will
+        # halt in windows waiting for a yes/no response to the unknown
+        # git host. this is normally handled by specifying
+        # "-o StrictHostKeyChecking=no" in the GIT_SSH executable, but it is
+        # still a mystery why this doesn't work properly in windows.
+        # so make a ssh call which creates the proper "known_hosts" file.
+        system("ssh -o StrictHostKeyChecking=no #{repo.url.split(':').first} exit 2>&1")
+      end
       return ''
     end
 
@@ -205,12 +211,13 @@ module RightScale
       git_cmd = "#{ssh_cmd} git fetch --tags --depth 1 2>&1"
       res = @watcher.launch_and_watch(git_cmd, @current_repo_dir)
       handle_watcher_result(res, 'git fetch', update, ok_codes=[0, 1]) # git fetch returns 1 when there is nothing to fetch
-      is_tag = is_branch = false
+      is_tag = is_branch = on_branch = false
       if succeeded? && (!update || @incremental)
         is_tag = `git tag`.split("\n").include?(@repo.tag)
         is_branch = `git branch -r`.split("\n").map { |t| t.strip }.include?("origin/#{@repo.tag}")
+        on_branch = is_branch && !!(`git branch`.split("\n").include?("* #{@repo.tag}")
       end
-      { :tag => is_tag, :branch => is_branch }
+      { :tag => is_tag, :branch => is_branch, :on_branch => on_branch }
     end
 
   end
